backend/requirements.txt
# Core data processing and analysis
pandas
numpy

# Yahoo Finance data provider
yfinance

# Visualization
matplotlib

# HTTP requests for OpenRouter API
requests
httpx[http2]
gunicorn
# Environment variables
python-dotenv

# Web server for API
Flask
Flask-Cors
